+++ src/transaction.ts
@@ -122,10 +122,12 @@ export type IAttributes<TStartParams, TResult, TError, TPotentialActions> = TSta
     tasksExportAction?: ITasksExportAction;
     /**
      * タスクエクスポート日時
+     * @deprecated use tasksExportAction
      */
     tasksExportedAt?: Date;
     /**
      * タスクエクスポート状態
+     * @deprecated use tasksExportAction
      */
     tasksExportationStatus: TransactionTasksExportationStatus;
     /**
@@ -158,8 +160,12 @@ export interface ISearchConditions<T extends TransactionType> {
     ids?: string[];
     /**
      * ステータスリスト
+     * @deprecated use tasksExportAction
      */
     statuses?: TransactionStatusType[];
+    status?: {
+        $in?: TransactionStatusType[];
+    };
     /**
      * 開始日時(から)
      */
@@ -177,16 +183,19 @@ export interface ISearchConditions<T extends TransactionType> {
      */
     endThrough?: Date;
     agent?: {
-        typeOf?: PersonType;
+        typeOf?: PersonType.Person | IWebApplication['typeOf'];
         ids?: string[];
         identifiers?: IIdentifier;
-        givenName?: string;
-        familyName?: string;
-        telephone?: string;
-        email?: string;
+        // givenName?: string; // discontinue(2024-06-17~)
+        // familyName?: string; // discontinue(2024-06-17~)
+        // telephone?: string; // discontinue(2024-06-17~)
+        // email?: string; // discontinue(2024-06-17~)
     };
-    tasksExportationStatuses?: TransactionTasksExportationStatus[];
+    // tasksExportationStatuses?: TransactionTasksExportationStatus[]; // discontinue(2024-06-17~)
     tasksExportAction?: {
-        actionStatus?: { $eq?: ActionStatusType };
+        actionStatus?: {
+            $eq?: ActionStatusType;
+            $in?: ActionStatusType[];
+        };
     };
 }